Example of a LoraWAN message.
We copied the LoRaWAN_APP files as we wanted to have something /without/
display in them. Our intent is to be able to use this code without use
of the display as that would take up a lot of power.

We're also considering removing the use of device_state, although that
would likely be a lot more work. Ideally, we want to encapsulate all
LoRaWAN things inside a class and not have all the global variables in
the main file.